[policies] Directly use a Transport for remote commands
Previously, remote command executions handled by policies were done by
moodifying the command string based on the `remote_exec` property of the
given `SoSTransport` in use for the node that the policy was loaded for.

While this worked well for SSH connections, newer transports may need to
do some manipulation of returned data in order for the rest of `sos
collect` to function as intended.

As such, switch to directly using a transport's `run_command()` method,
which will ideally handle any needed manipulations of either how the
command is execute and/or how the returned data is presented to the
calling component.

Related: #3087

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>