[build] Convert to setuptools
In python 3.12 distutils will be removed. As such, we need to update to
the replacement `setuptools`.

This commit makes the basic change over in `setup.py`, so that an
`sdist` source tarball can be generated. Note that while this source
tarball will still have the `.po` files in it any build tarball
(`bdist`) produced via the new `setup.py` will *not* have `.mo`
translation files compiled and included at this point.

In reviewing this change, it was found that our internationalization is
currently broken and very out of date. Future work will focus on fixing
that situation, but for now the immediate packaging needs are being
addressed.

Resolves: #3093

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>