[Policy] Add logging mechanism to Policy
This commit allows `Policy` to inherit the logging streams used by the
rest of sos. With this commit, this is primarily the `ui_log` in order
to report info and error level messages to the console. As was
discovered earlier, avocado and potentially other automation may
sometimes miss `print()` output, and so we should consistently use the
logging mechanism.

Future commits may be expected to enhace debug logging of Policy
functionality.

See also: #3106

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>