Remove duplicate commands
When a list of command are sent to run on a container the output
is printed to file on sos_commands/<plugin> folder, but also a
symbolic link is created from the sos_container/<container> file
to the sos_commands/<plugin> file. When a command is duplicated,
we get an exception on the plugin and the rest of commands are not
executed.

This patch removes some hardcoded commands, and clean the list from
duplicated commands before sending to the container.