Avoid overwrite symbolic link for a command output
Every output of a command run on a container is saved
in the sos_commands folder, also additionally a symbolic
link is created from the sos_container folder to link the
container where it was run to the output file. When there
is a duplicated command on the list to be run on the
container, the symbolic link creation will fail due to
file already exists error.

To avoid this error we can trust on the check_path method,
that will give us a clue avoid if the file already exists or
not to continue with the symbolic link creation.

Signed-off-by: Fernando Royo <froyo@redhat.com>