[report] Allow users to constrain sos process priority
Adds a new `--low-priority` option to report, which will attempt to
constrain the process priority for the report generation. We do this by
attempting to set ourselves to an 'idle' IO class, as well as setting
our niceness to 19 to avoid contending for CPU time.

This is also exposed via `sos collect`, however users should note that
this will not be effective until the sos-4.5.1 release.

Closes: #3127

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>