[tests] Add version consistency test
Commits a downstream test from RHEL to ensure that a given release has a
consistent version string in the UI report, manifest, and most
importantly for the package nvr for an sos package built from the
current branch. Note that this test is only implemented for RPM
installations at this time.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>