[component] only update handlers if preset with verbosity is used
Adding name for all handlers, with that they can be found by name
later and updated.

Also adding update=False parameter for _setup_logging method, with
that we can set it in component.py and only update it from
report/__init_.py if needed.

Lastly adding ifsfor creation of handlers and adding them to
logger instance, with that they should be defined just once and
we can only update their parameters instead of recreation.

Resolves: #3149

Signed-off-by: Jan Jansky <jjansky@redhat.com>