[ovn_central] add command ovn-appctl debug/chassis-features-list
The upcoming u/s OVN release will introduce a new command called
'debug/chassis-features-list', this patch add to the ovn_central
plugin's list of commands, in order to promptly obtain this
information.

Note that in case the OVN version doesn't support this command,
the plugin will not crash but instead save as output 'not a valid
command' msg.

Also a small refactor has been carried out in terms of reuse, for
function _find_sock, as ovn-appctl commands for northd service
need to run over specific socket.

Signed-off-by: Fernando Royo <froyo@redhat.com>