[tests] Allow tests to restrict execution architecture
Adds a new `arch` class attr that allows test classes to specify which
architecture(s) they are allowed/designed to execute on. If the test is
attempted to be executed on an architecture that is not in the specified
list, we will skip the test before any setup or sos executions are done.

By default, the `arch` class attr is an empty list, which implies that
the test can run on _any_ architecture. Tests should specify `arch` as a
list even for single-arch tests.

Included in this change is restricting the Foreman tests to x86_64 only,
as that is the only architecture on which that product is supported.

Closes: #3186

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>