[package_managers] Add MultiPackageManager support for policies
Adds a new `MultiPackageManager` class to handle systems that have
multiple distinct package managers on them.

This initial implementation requires policies that will leverage this to
define a `primary` package manager, and at least one fallback package
manager. Policies will need to pass the uninstantiated class object to
the initializing call to `MultiPackageManager`, for example:

    self.pkgman = MultiPackageManager(primary=RpmPackageManager,
                                      fallbacks=[DpkgPackageManager])

When a call to a `PackageManager` method is made to
`MultiPackageManager`, the call is first directed to the primary
manager. If that response is empty, None, etc... then the call is
forwarded to each of the fallback managers in the order they were
defined, until either a non-empty response is returned or we run out of
fallbacks to try.

In order to allow for some introspection of which defined manager
provides a non-empty response for package queries, a new `pkg_manager`
entry in the contents of the `packages` property has been added.

Closes: #2237
Related: #3214

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>