[package_managers] Break up `_generate_pkg_list()`
This commit breaks up `_generate_pkg_list()` by making it call to a new
`_parse_pkg_list()` method that each `PackageManager` will now need to
implement. This new method will be fed the output of the manager's
`query_command` for it to parse and yield package information, for
`_generate_pkg_list()` to then assemble into the `_packages` dict.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>