[ocp] Handle delete project timeouts more gracefully
In scenarios where the cluster is unable to delete the temporary project
due to a failure on one of the cluster nodes during the collection, we
would previously fail to delete the temporary project with a nasty
traceback after waiting for 5 minutes.

Be better at this by reducing the wait time (by adding a new `timeout`
parameter to `exec_primary_cmd()`) and more gracefully reporting the
error, and the fact that the user must make manual intervention to
remove the temporary project.

Related: RH SUPDEV-74

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>