[report] Don't use sysroot for network device enumeration
If sos is being used in a live environment to diagnose an issue, using
sysroot can cause the network device enumeration via /sys/class/net
crawling to fail. This will be the case for systems that do not use
`nmcli`.

When in a live environment, network devices will not be under
`/$sysroot/sys/class/net` but the "regular" path for the booted
environment. Similarly, if sos is being run in a container that is
properly configured, network devices will appear under `/sys/class/net`
and not (necessarily) under the sysroot path that mounts the host's
filesystem.

As such, disregard a configured sysroot when enumerating network devices
by crawling `/sys/class/net`, and trap any exceptions that may percolate
up from this in edge case environments.

Related: #3307

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>