[openvswitch] Add driverctl verbose option
driverctl commands give info about which drivers are used
for devices. They list PCI address and driver used e.g.

0000:01:00.0 vfio-pci

This can usually be correlated to the device type through lspci.
However, lspci may not be available if only some plugins were
captured as part of the report.

Even when lspci info is available, it is still convienent to have
the device type co-located with the driver and override information.

Add verbose option to driverctl commands that support it to also
list device type. e.g. (wrapped for line length)

0000:01:00.0 vfio-pci
 (82599ES 10-Gigabit SFI/SFP+ Network Connection
 (Ethernet 10G 4P X520/I350 rNDC))

Signed-off-by: Kevin Traynor <ktraynor@redhat.com>