[debian] Packaging and testing for debian pkgs
Initial start of building and testing the deb package

This will improve the overall testing of sos over the Ubuntu releases,
and also helps with the basis of thenadding further stagetwo testing in
the future. This will help to pick up issues automatically rather than
the manual testing.

Add ubuntu-latest for deb CI build and testing. Update the GCE images
for all ubuntu images for CI testing

Co-authored-by: Nikhil Kshirsagar <nikhil.kshirsagar@canonical.com>
Signed-off-by: Nikhil Kshirsagar <nikhil.kshirsagar@canonical.com>
Signed-off-by: Arif Ali <arif.ali@canonical.com>