[arcconf] add getconfig and GETLOGS commands for all available controllers
Currently, the arcconf getconfig and GETLOGS commands are added with the
assumption that the system has only one controller with ID 1. It is
possible that a system can have multiple controllers, so commands need
to be added to extract information for all available controllers.

The arcconf utility does not have a dedicated command to get the list of
controllers available in the system. Therefore, parse the 'arcconf list'
output to obtain the controller ID for all controllers in the system,
and add getconfig and GETLOGS commands for each available controllers.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>
Suggested-by: Borislav Stoymirski <borislav.stoymirski@bg.ibm.com>