[collect] update for strict confinement for juju
With juju versions 3 and above, when collecting the tarballs
from machines it will grab them into a strictly confined area.
This means that we need to be able to grab the data to a different
location, such as somewhere in the home directory. Using a directory
such as ~/.cache/sos-collect-juju is being used to overcome this
issue.

Related: #3399
Signed-off-by: Arif Ali <arif.ali@canonical.com>