[collect] update for strict confinement for juju
With juju versions 3 and above, when collecting the tarballs
from machines it will grab them into a strictly confined area.
This means that we need to be able to access this area via sudo.

In order for this now to be fully supported, we need sudo on the
host that is running juju, otherwise sos collect on a juju
environment will not work.

Related: #3399
Signed-off-by: Arif Ali <arif.ali@canonical.com>