[policies] Fix dist_version for Ubuntu
DISTRIB_RELEASE variable contains a float value for the Ubuntu
release version and not an integer value.

Example, a 22.04 release is an LTS release, and a 22.10 release
is an interim non-LTS release, changing this value to a float
allows us to retrieve the full version rather than just the int.

Signed-off-by: David Negreira <david.negreira@canonical.com>