[global] Replace use of `.format()` with f-strings
As part of the project's effort to modernize on f-strings, this commit
replaces all usage of .format() with f-string equivalents. The vast
majority of these are in-place syntax changes, but a limited number
change the order of list items or some formatting tricks in order to
appease PEP8.

None of these conversions change the underlying logic of the flows they
appear in.

Related: #3472 discussion

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>