[lxd] Fix predicate for lxc commands collection for snap
The list of modules was too broad, and depending on the different attributes
configured the list of modules to be loaded could also be different.
Liaising with the LXD team suggested to check for the `snap.lxd.daemon`
service rather then the modules to collect `lxd.buginfo`. This is due to the
fact that LXD is socket activated, and running any lxd based command could
start the service and hence load the modules. Gating this with either the
`snap.lxd.daemon` for snap or `lxd` service for debs respectively should
prevent this from happening.

This also no obfuscates the certificate that is retrieved from `lxd.buginfo`

Resolves: SET-377

Signed-off-by: Arif Ali <arif.ali@canonical.com>