[ceph] Add Reef release commands
Ceph release 18.2.0 added a couple of new commands 'counter dump'
and 'counter schema' and will supersede 'perf dump' and 'perf schema'
which are deprecated now. For sosreport, we obviously need both
commands to collect data on existing deployments.

At some point in the future, we can remove 'perf dump' and
'perf schema' from the list of commands for Ceph plugins.

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>