[collect] refactor _format_version
Ensure that we format the package versions to the pep440 standard.
This is necessary as we are using `parse_version()` from the `packaging`
package to verify that the sos collector node version is equal or bigger
than the versions we are collecting from the nodes. If we pass the wrong
format to `parse_version()`, we are unable to do the comparison and
unable to run `sos report` on those nodes.
This addresses the issue of the Ubuntu packaging version naming with `+`
and `~` signals as well as generic versioning schemes.

Resolves: #3544

Signed-off-by: David Negreira <david.negreira@canonical.com>