[SCLPlugin] Drop SCLPlugin support
SCL is not used in RHEL past RHEL 7, and sos has not directly supported
RHEL 7 for some years now due to the switch to python3 that came along
with sos-4.0.

Drop SCLPlugin, as it will never be used in modern sos. Cleanup any
references to it in existing plugins as well.

Resolves: #3559

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>