[convert2rhel] Update copy spec to include * report
convert2rhel now generates a post-conversion report as well as the
pre-conversion report. We want to be able to collect both reports during
sosreport execution. To make sure we are not missing them (or any other
future report we may introduce), the copy spec string was changed to use
a `*` instead of the names of the reports.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>