[global] Remove empty except warnings highlighted by CodeQL
This commit is a first pass at addressing a number of the `Empty Except`
warnings flagged by CodeQL. These are for try/except blocks whose except
statement simply contains a `pass`, and to address these we add a log
message in almost all cases. Where we don't log, there is an added
comment for context as to why we're passing without further action.

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>