[tox] Use posargs for the tests
This allows to run specific tests, and not all the tests in one go.
This is especially useful if you're writing a plugin or creating a test
for a specific scenario. Below an example of what would be possible

```
sudo tox -e stagetwo_tests -- tests/report_tests/plugin_tests/openstack/openstack.py
```

This will run stagetwo tests for the specific test, i.e. openstack in
this case

This also fixes the stagetwo tests, so that it picks up the systemd
python libs from the distro environment

Update the installation of avocado-framework from one place, so that we
don't need to change the version in many places. Adding
`test-requirements.txt` to handle this and integrate this with both tox
and Cirrus CI

Signed-off-by: Arif Ali <arif.ali@canonical.com>