[apt] Obfuscate passwords for configs/sources
The configuration of apt was not being obfuscated if any of the URLs
had a password; one example was the proxy. There is also a new format
for the deb sources, so ensuring that would be obfuscated in the same
way. Add some tests to ensure that this is always being obfuscated.

Resolves: SET-732

Signed-off-by: Arif Ali <arif.ali@canonical.com>