[kernel] Add plain 'dmesg' command to plugin
PR #3657 added human readable timestamps to the dmesg
command, but the command can be unreliable. This PR
adds the plain 'dmesg' command capture again, to make
sure that we still have an output with reliable
timestamps that can be easily converted to human readable
in these cases where 'dmesg -T' could not be trusted.

Signed-off-by: Jose Castillo <jcastillo@redhat.com>