[Plugin] Standardize directory listings with new method
This commit adds a new `add_dir_listing()` wrapper method to standardize
the collection of directory content listings across plugins. By default
this will use the `ls` command, and will use `tree` if the `tree`
parameter is set to `True`.

This wrapper is not intended to replace directory listing that are
handlded by other components with their own `ls` functionality (e.g.
`npm ls` in the `npm` plugin). Additionally this does not replace the
direct calling of `ls` when used with `Plugin.exec_cmd()` to take some
action based on the output of the `ls` command.

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>