[Global] Update strings to align with `sos` tooling names
This commit serves to reinforce the typing of "sos (space) report" in
place of the legacy "sosreport", and similarly updates items such as
"sos-collector" to "sos collect". The tool is `sos` ("ess-oh-ess"), of
which `report` and `collect` are sub-commands.

The most visible change is the update to the line printed before the
path to a generate sos report, which may be significant for scripts
and/or automation. To be clear, this commit changes this line:

`Your sosreport has been generated and saved in:`

to

`Your sos report has been generated and saved in:`

Note that this does not change the use of `sosreport` when used as part
of method names, nor does it change the *filename* of the generated
archive for either `report` or `collect`.

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>