[clean,options] Correct skip-clean-files in sos.conf and skipping logic
First, fix the issue raised in #3671 for the `--skip-cleaning-files`
option, which would need to be written to `sos.conf` as
`skip-clean-files` due to the internal variable naming.

Second, noticed as part of the reproducer for this issue when
`--skip-cleaning-files` would be seti, that we were inadvertantly still
going through the motions of creating a temporary file and scanning the
file even if we weren't intending to do any obfuscations to it. Correct
this by bailing out after determining there are no parsers for which we
want to do an obfuscation. Apply the same update to symlink paths so
that we don't create broken symlinks.

Resolves: #3671

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>