[PluginOpt] Support different parameter name
Current `PluginOpt` code doesn't have the possibility of having a
different name for the parameter and the reference used in the code.

This would be the equivalent in arg parser to not having a `dest`
argument available to us.

This patch proposes a change to allow this using a new `PluginOpt` field
called `param_name`.

This new field will only be used when interfacing with users. In other
words, when showing the options to users (eg: `list_plugins`), when
parsing the command line arguments (eg: `_set_tunables`), and when
reading the config file (this doesn't require code changes because it
leverages the normal `_set_tunables` method).

Signed-off-by: Gorka Eguileor <geguileo@redhat.com>