[sos] Add 'upload' component to upload existing reports and files
This commit marks the beginning of the addition of a new 'upload'
component for sos, which can be used to upload already created
sos reports, collects, or other files like logs or vmcores to
a policy defined location.

The user needs to specify a file location, and can make use of any
of the options that exist nowadays for the --upload option.

This first commit includes:
- The initial framework for the 'upload' component.
- The new man page for 'sos upload'.
- The code in the component 'help' to show information about
the component.
- The code in sos/__init__.py to deal with the component.
- The code for uploads to Red Hat and Ubuntu systems.
- The code to allow uploads specifying remote destination, called
targets in this implementation. For example, you could generate
a sos report in a CentOS system and specify a target defined as
'redhat' or 'RedHatUpload' to upload to the Red Hat Customer Portal.
- And modifications to setup.py to build the man pages.

Related: RHEL-23032, SUPDEV-138, CLIOT-481

Co-authored-by: Jose Castillo <jcastillo@redhat.com>
Co-authored-by: Pavel Moravec <pmoravec@redhat.com>
Co-authored-by: Trevor Benson <trevor.benson@gmail.com>
Signed-off-by: Jose Castillo <jcastillo@redhat.com>