[tests] Allow tests to restrict OS version
Add a new `versions` class attribute that allows test classes
to specify which version(s) are allowed/designed to execute on.
If the test is attempted to be executed on a version that's not
specified, the test will be skipped and a message will be printed.

By default, the list of versions will be empty, which means that
the test should run in all versions of the OS.

This patch contains a change to restrict the execution of Teamd
to RHEL 8, because it was deprecated for RHEL 9 and later.

Signed-off-by: Jose Castillo <jcastillo@redhat.com>