[ssh] Stop collecting 'ls' information from non-local home dirs
Commands like 'ls -laZ' can take some time to run in remote filesystems,
for example NFS. When you have hundreds of them, sos can take a long
time to walk through it, and could be even worse if we use the option
'recursive'.
For ssh home dirs, we don't need the listing, so in this patch we detect
the type of filesystem and run the command only on local ones.

Related: RHEL-22389, SUPDEV-156

Signed-off-by: Jose Castillo <jcastillo@redhat.com>