[Plugin] Check if a path exists inside a container in add_dir_listing()
Currently, add_dir_listing() checks if a path exists outside of a
container, and when we call the command for a container, it fails to
collect any listing. This patch adds a new function
container_path_exists() that checks for the specified path in
add_dir_listing() inside the container.

Signed-off-by: Jose Castillo <jcastillo@redhat.com>