[collect] Fix exception when using collect with batch and password
Using --batch --password together in sos-collect leads to an error:

Unable to open remote session:
unsupported operand type(s) for +: 'bool' and 'str'

This commit adds a check for both options running at
the same time and if so, prints a message and exits.
Also, the commit changes the man page to reflect
this restriction.

Related: #3839

Signed-off-by: Jose Castillo <jcastillo@redhat.com>