[upload] Add new component
This commit marks the beginning of a series to implement the upload
functionality of collect and report into a separate component. This new
upload component is intended to be primarily used to define standard
"targets" for uploading support-relevant files to support-providing
vendors, e.g. Red Hat and Canonical.

Both `report` and `collect` will be made to leverage this component in
later commits to this series.

As a practical example, the goal for the end-user experience is a
command like the following for providing data to a support vendor:

  + Uploading in-line with report (or collect archive) creation
    `sos report --case-id 123456 --upload --upload-target=redhat`
  + Uploading a report after generation
    `sos upload sosreport-example.tar.xz --target=canonical`
  + Uploading non-sos data that is relevant to an ongoing support case
    `sos upload $some_helpful_file --target=my_awesome_vendor`

Additionally, there will still be support for "generic" targets such as
FTP or SFTP destinations that users may wish to define themselves, and
still allow for automatic transfer of generate archives to.

This first commit provides the framework for the component directly, as
well as the "upload target" abstraction which serves as an entrypoint
for standardized destinations for uploads.

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>