[upload] Add HTTP(S) and Canonical upload targets
Adds two new upload targets - a generic http(s) target and a port of the
existing ubuntu policy upload to a `canonical` upload target.

Note that this also changes the `upload-method` option to
`--upload-http-method` and also removes `auto` as a valid choice,
leaving just `put` and `post`, as it is assumed that the `auto` logic
would now be redundant for independently implemented vendor targets that
subclass the generic http target.

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>