[utilities] Add tac_logs
tac_logs() reverses the order of the logs in a text
file but keeps multiline logs in order.
It is aimed at reversing 'journalctl --reverse'.

This avoids depending on GNU coreutils version of tac
(busybox tac is missing some flags) and is way faster.
Testing with 100MB of logs:
- tac -brs '^[^ ]' (handle multiline): ~30s
- tac_logs (our python implementation): ~0.7s
- tac (not handling multiline logs): ~0.3s

If we are concerned about disk usage, we could call
mmap.resize() to truncate the input file while processing.

Signed-off-by: Etienne Champetier <e.champetier@ateme.com>