[Plugin,crio,gluster] Refactor signalling processes to collect stackdumps
Adds a new `Plugin.signal_process_usr()` method that standardizes the
way we send SIGUSR1 to process(es) by name. This new method will return
a list of signalled pids for plugins to iterate over if needed to
determine further collections.

Updates the `crio` and `gluster` plugins to leverage this new method in
place of their previous approaches.

Signed-off-by: Jake Hunsaker <jacob.r.hunsaker@gmail.com>