[logs] Add all auth.log, syslog and kerne.log as standard
The RH based systems already collect this type of logs, and Ubuntu systems was
lacking this. This will allow to debug issue better and easier to consume.

Re-order the file set to be in alphabetical order.

The else for the syslog, auth.log and syslog us now redundant, as we collect all
of these files now.

Closes: #3940

Signed-off-by: Arif Ali <arif.ali@canonical.com>