[squid] updated Debian and Ubuntu paths in Squid plugin
The Squid plugin was using squid3 paths for all Debian
and Ubuntu versions, but this path has not been correct
since Debian 9 Stretch and Ubuntu 18.04 Bionic Beaver.
The result is that no Squid data is recorded in current
OS releases.

This fixes the paths for both distros and cleans up the
self.add_copy_spec() line to be consistent with style
guidelines.

Signed-off-by: Daniel Emmons <dan.emmons@canonical.com>