[memory] Collect all sysfs THP files
Changes to THP result in sysfs `enabled` file alone (and the kernel
parameter) are now insuffiecient to know if THP is enabled or not. For
example, `shmem_enabled` can be set to `[always]` while `enabled` can
be set to `[never]` and THP will still be enabled but only for shmem
segments. As such, to better know if and how THP is enabled, collect
all files under `/sys/kernel/mm/transparent_hugepage/` rather than
just `enabled`.

Signed-off-by: Charles Haithcock <chaithco@redhat.com>