[login,clean] Update login plugin and use for clean
Add lslogin and lastlog2 output. Then use these for username
obfuscations for clean. New distros don't have lastlogin anymore and
last is not always installed by default.

timout: is excluded as well. When the command is run, it runs with
timout command, and if the command doesn't exist, it parses that as
the username.

Related: #3960

Signed-off-by: Arif Ali <arif.ali@canonical.com>