[tests] Fix issue with 127.0.1.1 IP being detected
Running stageone tests in some environments sees the hostname of the
machine to have an IP of 127.0.1.1 but the host is never set up in a
way that this address is configured on the host. We use socket.connect
to connect to a dummy address, which ensure to go via a default route
and therefore not track the loopbak address. Then grabbing the IP of
the socket give you a proper local IP address.

Closes: #3975
Signed-off-by: Arif Ali <arif.ali@canonical.com>