[tests] Fix issue with 127.0.1.1 IP being detected
Running stageone tests in some environments sees the hostname of the
machine to have an IP of 127.0.1.1 but the host is never set up in a
way that this address is configured on the host. We check for
hostname, if this returns a loopback address then we add .local to
the end of the hostname, which ensures we get the address of the host.

Closes: #3975
Signed-off-by: Arif Ali <arif.ali@canonical.com>