[dqlite] Conduct Dqlite Collection in Consumers
Collect files relating to the configuration of the distributed cluster
that underlies micro* applications like microceph, microovn, microcloud,
microk8s, and lxd. Collect the directory listing of the dqlite database
dir to check for corrupt intervals, and collect the expiry date of
cluster certs. As well, inspect information pertinent to the cluster
by querying against the dqlite db.

Signed-off-by: Bryan Fraschetti <bryan.fraschetti@canonical.com>